Fix regex mistake
The regex to normalise an ID mistakenly required a leading "/" to
"anchor" the ID as part of a URL. This of course fails when presented
with both a pure ID suffix and various (nonstandard but plausible) other
input strings.

Therefore we remove the "/" from the regex and just search for the ID
suffix pattern and return the first result. Note that this change is
"safe" in that it is "backwards-compatible" while gaining the ability to
correctly match more input strings.